split kibana files to read and payload
